cpu: pbtb, fixed vanilla pred update in fetch
Previously: we'd sometimes use the predictor for a pb, but not use that
prediction because e.g. the pbtb already had a definitive answer,
and so the incorrect vanilla prediction would never be amended because
that pb wouldn't squash.

Fixed (i think) it to squash the vanilla predictor whenever it
diverges from what we end up using in fetch.

Change-Id: If512bb991316ceb37bacfe94e6c680bc1a2af6c4